documentation-crawler: Create common spider for documentation checking.
Fixes #2639
